+++ src/config/webpack.ts
@@ -32,6 +32,7 @@ export default {
 				path: `${dest}/server`,
 				filename: '[name].js',
 				chunkFilename: '[hash]/[name].[id].js',
+				publicPath: `client/`,
 				libraryTarget: 'commonjs2'
 			};
 		}