+++ src/core/create_compilers/RollupCompiler.ts
@@ -1,7 +1,11 @@
 import * as path from 'path';
+import color from 'kleur';
 import relative from 'require-relative';
+import { RollupError } from 'rollup/types';
 import { CompileResult } from './interfaces';
 import RollupResult from './RollupResult';
+
+const stderr = console.error.bind(console);
 
 let rollup: any;
 
@@ -37,7 +41,7 @@ export default class RollupCompiler {
 			transform: (code: string, id: string) => {
 				if (/\.css$/.test(id)) {
 					this.css_files.push({ id, code });
-					return ``;
+					return {code: '', moduleSideEffects: 'no-treeshake'};
 				}
 			}
 		});
@@ -71,16 +75,10 @@ export default class RollupCompiler {
 
 			return new RollupResult(Date.now() - start, this, sourcemap);
 		} catch (err) {
-			if (err.filename) {
-				// TODO this is a bit messy. Also, can
-				// Rollup emit other kinds of error?
-				err.message = [
-					`Failed to build — error in ${err.filename}: ${err.message}`,
-					err.frame
-				].filter(Boolean).join('\n');
-			}
+			// flush warnings
+			stderr(new RollupResult(Date.now() - start, this, sourcemap).print());
 
-			throw err;
+			handleError(err);
 		}
 	}
 
@@ -150,8 +148,12 @@ export default class RollupCompiler {
 			}
 		});
 
-		const resp = await bundle.generate({ format: 'cjs' });
-		const { code } = resp.output ? resp.output[0] : resp;
+		const {
+			output: [{ code }]
+		} = await bundle.generate({
+			exports: 'named',
+			format: 'cjs'
+		});
 
 		// temporarily override require
 		const defaultLoader = require.extensions['.js'];
@@ -163,9 +165,46 @@ export default class RollupCompiler {
 			}
 		};
 
-		const config: any = require(input);
+		const config: any = require(input).default;
 		delete require.cache[input];
 
 		return config;
 	}
+}
+
+
+// copied from https://github.com/rollup/rollup/blob/master/cli/logging.ts
+// and updated so that it will compile here
+
+export function handleError(err: RollupError, recover = false) {
+	let description = err.message || err;
+	if (err.name) description = `${err.name}: ${description}`;
+	const message =
+		(err.plugin
+			? `(plugin ${(err).plugin}) ${description}`
+			: description) || err;
+
+	stderr(color.bold().red(`[!] ${color.bold(message.toString())}`));
+
+	if (err.url) {
+		stderr(color.cyan(err.url));
+	}
+
+	if (err.loc) {
+		stderr(`${(err.loc.file || err.id)!} (${err.loc.line}:${err.loc.column})`);
+	} else if (err.id) {
+		stderr(err.id);
+	}
+
+	if (err.frame) {
+		stderr(color.dim(err.frame));
+	}
+
+	if (err.stack) {
+		stderr(color.dim(err.stack));
+	}
+
+	stderr('');
+
+	if (!recover) process.exit(1);
 }