+++ src/core/create_manifest_data.ts
@@ -1,6 +1,5 @@
 import * as fs from 'fs';
 import * as path from 'path';
-import svelte from 'svelte/compiler';
 import { Page, PageComponent, ServerRoute, ManifestData } from '../interfaces';
 import { posixify, reserved_words } from '../utils';
 
@@ -13,19 +12,6 @@ export default function create_manifest_data(cwd: string, extensions: string = '
 		throw new Error(`As of Sapper 0.21, the routes/ directory should become src/routes/`);
 	}
 
-	function has_preload(file: string) {
-		const source = fs.readFileSync(path.join(cwd, file), 'utf-8');
-
-		if (/preload/.test(source)) {
-			try {
-				const { vars } = svelte.compile(source.replace(/<style\b[^>]*>[^]*?<\/style>/g, ''), { generate: false });
-				return vars.some((variable: any) => variable.module && variable.export_name === 'preload');
-			} catch (err) {}
-		}
-
-		return false;
-	}
-
 	function find_layout(file_name: string, component_name: string, dir: string = '') {
 		const ext = component_extensions.find((ext) => fs.existsSync(path.join(cwd, dir, `${file_name}${ext}`)));
 		const file = posixify(path.join(dir, `${file_name}${ext}`))
@@ -33,8 +19,7 @@ export default function create_manifest_data(cwd: string, extensions: string = '
 		return ext
 			? {
 				name: component_name,
-				file: file,
-				has_preload: has_preload(file)
+				file: file
 			}
 			: null;
 	}
@@ -47,16 +32,14 @@ export default function create_manifest_data(cwd: string, extensions: string = '
 		default: true,
 		type: 'layout',
 		name: '_default_layout',
-		file: null,
-		has_preload: false
+		file: null
 	};
 
 	const default_error: PageComponent = {
 		default: true,
 		type: 'error',
 		name: '_default_error',
-		file: null,
-		has_preload: false
+		file: null
 	};
 
 	function walk(
@@ -84,16 +67,16 @@ export default function create_manifest_data(cwd: string, extensions: string = '
 					? basename
 					: basename.slice(0, -ext.length);
 
+				if (/\]\[/.test(segment)) {
+					throw new Error(`Invalid route ${file} — parameters must be separated`);
+				}
+
 				const parts = get_parts(segment);
 				const is_index = is_dir ? false : basename.startsWith('index.');
 				const is_page = component_extensions.indexOf(ext) !== -1;
 				const route_suffix = basename.slice(basename.indexOf('.'), -ext.length);
 
 				parts.forEach(part => {
-					if (/\]\[/.test(part.content)) {
-						throw new Error(`Invalid route ${file} — parameters must be separated`);
-					}
-
 					if (part.qualifier && /[\(\)\?\:]/.test(part.qualifier.slice(1, -1))) {
 						throw new Error(`Invalid route ${file} — cannot use (, ), ? or : in route qualifiers`);
 					}
@@ -161,8 +144,7 @@ export default function create_manifest_data(cwd: string, extensions: string = '
 			else if (item.is_page) {
 				const component = {
 					name: get_slug(item.file),
-					file: item.file,
-					has_preload: has_preload(item.file)
+					file: item.file
 				};
 
 				components.push(component);
@@ -298,7 +280,7 @@ function comparator(
 }
 
 function get_parts(part: string): Part[] {
-	return part.split(/\[(.+)\]/)
+	return part.split(/\[(.+?\(.+?\)|.+?)\]/)
 		.map((str, i) => {
 			if (!str) return null;
 			const dynamic = i % 2 === 1;
@@ -340,7 +322,8 @@ function get_pattern(segments: Part[][], add_trailing_slash: boolean) {
 					.replace(/\?/g, '%3F')
 					.replace(/#/g, '%23')
 					.replace(/%5B/g, '[')
-					.replace(/%5D/g, ']');
+					.replace(/%5D/g, ']')
+					.replace(/[.*+?^${}()|[\]\\]/g, '\\$&');
 		}).join('');
 	}).join('\\/');