+++ src/utils.ts
@@ -117,3 +117,9 @@ export const reserved_words = new Set([
 	'with',
 	'yield',
 ]);
+
+export function normalize_path(user_path) {
+	const p = path.normalize(user_path);
+	// normalize drive letter on Windows
+	return p.length ? p.charAt(0).toLowerCase() + p.slice(1) : '';
+}