+++ test/apps/credentials/src/routes/credentials/authorization.json.js
@@ -1,0 +1,13 @@
+export function get(req, res) {
+	if (req.headers.authorization) {
+		res.writeHead(200);
+		res.end(JSON.stringify({
+			message: req.headers.authorization
+		}));
+	} else {
+		res.writeHead(200);
+		res.end(JSON.stringify({
+			message: 'unauthorized'
+		}));
+	}
+}