+++ test/apps/credentials/src/routes/credentials/cookie.json.js
@@ -3,19 +3,12 @@ import cookie from 'cookie';
 export function get(req, res) {
 	if (req.headers.cookie) {
 		const cookies = cookie.parse(req.headers.cookie);
-
-		res.writeHead(200, {
-			'Content-Type': 'application/json'
-		});
-
+		res.writeHead(200);
 		res.end(JSON.stringify({
 			message: `a: ${cookies.a}, b: ${cookies.b}, max-age: ${cookies['max-age']}`
 		}));
 	} else {
-		res.writeHead(403, {
-			'Content-Type': 'application/json'
-		});
-
+		res.writeHead(200);
 		res.end(JSON.stringify({
 			message: 'unauthorized'
 		}));