+++ test/apps/credentials/src/routes/index.svelte
@@ -1,0 +1,3 @@
+<h1>Great success!</h1>
+
+<a id="cookie" href="credentials?creds=include&amp;type=cookie">link</a>